monarch_tensor_worker

Struct WorkerActor

pub struct WorkerActor { /* private fields */ }

A PyTorch runtime instance, operating on a single accelerator device, controlled via hyperactor messaging.

Generally this is a thin multiplexer over a set of [Stream]s that do the real work.

See WorkerMessage for what it can do!